+++ composeApp/src/commonMain/kotlin/com.cheesecake.mafia/state/PlayerState.kt
@@ -4,11 +4,11 @@ import com.cheesecake.mafia.viewModel.Item
 
 @kotlinx.serialization.Serializable
 data class PlayerState(
-    val id: Int, val name: String
+    val id: Long, val name: String
 )
 
 @kotlinx.serialization.Serializable
-sealed class SelectPlayerState(val name: String, val id: Int) {
+sealed class SelectPlayerState(val name: String, val id: Long) {
     data object None: SelectPlayerState("", -1)
     data class New(val value: String): SelectPlayerState(value, 0)
     data class Exist(val player: PlayerState): SelectPlayerState(player.name, player.id)
